Removed global variables. Began refractoring of server.js, removing request calls for axios calls.
