+++ package.json
@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "^0.25.0",
+    "cookie-session": "^2.0.0",
     "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "request": "^2.88.2"
+    "express": "^4.17.1"
   },
   "scripts": {
     "start": "node server/server.js",
